test: cover pending push target rebinding
